requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python-headless
numpy
rich
